refactor(test): use production build instead of source
